+++ src/main/java/com/rtrnonato/library_management/entities/User.java
@@ -20,7 +20,7 @@ public class User implements Serializable {
 	
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Integer id;
+	private Long id;
 	private String name;
 	private String email;
 	
@@ -32,17 +32,17 @@ public class User implements Serializable {
 		
 	}
 	
-	public User(Integer id, String name, String email) {
+	public User(Long id, String name, String email) {
 		this.id = id;
 		this.name = name;
 		this.email = email;
 	}
 
-	public Integer getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
